refactor: 출력 객체에서 에러와 통계 출력 기능 분리
