+++ src/main/java/lotto/view/common/ErrorMessageView.java
@@ -1,0 +1,9 @@
+package lotto.view.common;
+
+public class ErrorMessageView {
+    private static final String ERROR_MESSAGE = "[ERROR] ";
+
+    public static void printError(IllegalArgumentException e) {
+        System.out.println(ERROR_MESSAGE + e.getMessage());
+    }
+}
